create hook for token-list provider
